+++ backend_test.ts
@@ -9,7 +9,7 @@ import axios from 'axios';
 import * as fs from 'fs';
 import * as path from 'path';
 
-const API_BASE = 'https://merchant-integration-3.preview.emergentagent.com/api';
+const API_BASE = 'https://merchant-checkout-22.preview.emergentagent.com/api';
 
 interface TestResult {
   test: string;